recommendations: buttons payment and cancel
Has it ever been recorded? Sorry.

The changes occurred because VScode organizes CSS properties alphabetically.
In fact, the changes were just the ones that are shown in this new commit.

Well, I don't want to get in the way. It was just a suggestion.
But I'm glad you answered. I'm on my JS learning journey and your posts on freecodecamp helped me a lot.
Thank you so much, really